requirements.txt
pandas==1.5.3
matplotlib
seaborn
numpy==1.23.5
scipy
streamlit
plotly
scikit-learn
openai
python-dotenv
openpyxl
kaleido
streamlit-option-menu
anthropic
google-generativeai
pandasai